requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==23.2.0
blessed==1.20.0
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.3.0
cycler==0.12.1
databricks-sdk==0.36.0
datasets==3.1.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
editor==1.6.6
fastapi==0.115.4
filelock==3.16.1
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.35.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphene==3.4.1
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.0.0
inquirer==3.4.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
litellm==1.51.3
-e git+ssh://git@github.com/slashml/magemaker.git@5dfe75ed5eb99114ed673b3d1e7b3ab81412ad5a#egg=magemaker
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mlflow==2.17.2
mlflow-skinny==2.17.2
mock==4.0.3
multidict==6.1.0
multiprocess==0.70.16
numpy==1.26.4
openai==1.53.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
packaging==24.1
pandas==2.2.3
pathos==0.3.2
pfzy==0.3.4
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pox==0.3.5
ppft==1.7.6.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
runs==1.2.2
s3transfer==0.10.3
safetensors==0.4.5
sagemaker==2.232.3
sagemaker-core==1.0.13
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.41.2
tblib==3.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
Werkzeug==3.1.1
wrapt==1.16.0
xmod==1.8.1
xxhash==3.5.0
yarl==1.17.1
zipp==3.20.2
